requirements.txt
pydantic-ai
openai
pypdf
lancedb
streamlit
python-dotenv
